Fix Part Two

+++ Plugins/datTest/datTestTypes.h
@@ -36,7 +36,7 @@ private:
   double gatherRadius = 20.0;
 
   //No flower can live longer than 60 days
-  unsigned int maxAge = 60*TICKSINADAY;
+  unsigned long long int maxAge = 60*TICKSINADAY;
   double chanceOfRandomDeath = 0.0005; //0.05% chance of dying at any point
 
   double chance;
